requirements.txt
ollama
langchain
langchain-text-splitters
faiss-cpu
unstructured
markdown
langchain-community
rich
playwright
html2text
beautifulsoup4
tldextract

chromadb
llama-index-vector-stores-chroma
llama-index-core
llama-index-readers-web
llama-index-llms-ollama
llama-index-embeddings-ollama
llama-index-llms-openai
